- Remove license fields from several dependencies in package-lock.json for cleaner output. - Update versions of multiple dependencies, including @emnapi/runtime, @floating-ui/core, and typescript for improved functionality and security. - Refactor ContactForm component to use useRouter from Next.js for navigation instead of useTranslations. - Simplify form submission handling by redirecting to a thank-you page upon successful submission. - Remove unnecessary loading state and related UI elements from ContactForm for a more streamlined user experience.
